refactor(apphosting): decouple compare dashboard into static public assets, add loading spinners and performance tiles
