+++ src/apphosting/compare/README.md
@@ -50,7 +50,69 @@ FIREBASE_CLI_EXPERIMENTS=apphosting firebase apphosting:compare-suite --project 
 FIREBASE_CLI_EXPERIMENTS=apphosting firebase apphosting:compare-suite --serve --port 3000
 ```
 
+
+## System Architecture & Subsystem Delineation
+
+The N-Way Matrix Comparison Tool is divided into two distinct subsystems to separate the CLI recording/deployment engine from the visual comparison dashboard (the Diff Viewer).
+
+
+```mermaid
+graph TD
+    %% Define styles
+    classDef diffViewer fill:#1e293b,stroke:#3b82f6,stroke-width:2px,color:#f8fafc;
+    classDef engine fill:#0f172a,stroke:#334155,stroke-width:1px,color:#94a3b8;
+    classDef ui fill:#1e1b4b,stroke:#818cf8,stroke-width:2px,color:#f8fafc;
+    classDef cache fill:#022c22,stroke:#10b981,stroke-width:2px,color:#f8fafc;
+
+    subgraph DiffViewerGroup["Diff Viewer Web Application (Visual Dashboard)"]
+        UI["User Interface (Web Dashboard UI)<br/>index.html / index.css / index.js"]:::ui
+        Server["Diff Viewer API Server<br/>server.ts"]:::diffViewer
+    end
+
+    subgraph EngineGroup["Deployment & Recording Engine (CLI Runner)"]
+        Cache["Cache Manager (JSON)<br/>cache.ts"]:::cache
+        Comp["Comparison Engine<br/>compare.ts"]:::engine
+        Crawler["Crawler & Route Discovery<br/>crawler.ts / discover.ts"]:::engine
+        Orch["Orchestration Engine<br/>suite.ts / lifecycle.ts"]:::engine
+    end
+
+    %% Relationships
+    UI -- "HTTP / API Requests" --> Server
+    Server -- "Read JSON Cache" --> Cache
+    Cache -- "Write Telemetry" --> Comp
+    Comp -- "Run Parity Analysis" --> Crawler
+    Crawler -- "Spider Dynamic Paths" --> Orch
+
+    class UI,Server diffViewer;
+    class Cache cache;
+    class Comp,Crawler,Orch engine;
+```
+
+
+### 1. The Diff Viewer Web Application (Visual Dashboard)
+This subsystem is responsible for hosting the interactive web interface, serving visualization assets, and exposing the matrix-calculation APIs.
+*   **`src/apphosting/compare/public/`**: Contains the frontend single-page application:
+    *   `index.html`: Layout structure, split panes, controls bar, and dialog anchors.
+    *   `index.css`: Stylesheet, custom dark-mode slate theme, responsive grid cards, and transition overlays.
+    *   `index.js`: Client-side logic, real-time matrix rendering using HSL interpolation, localStorage configuration persistence, inline table ignores, and dynamic filters.
+*   **`src/apphosting/compare/server.ts`**: The lightweight Express API backend. It reads recorded JSON files from the cache directory, serves the static assets from the `public/` directory, and provides endpoints:
+    *   `/api/recordings`: Lists all available cached test suite recordings.
+    *   `/api/matrix`: Recalculates pairwise parity similarity scores dynamically.
+    *   `/api/compare`: Performs exact body and header diffs between two selected variant recordings.
+    *   `/api/render`: Proxies the recorded response bodies as live visual rendering previews.
+
+### 2. The Deployment & Recording Engine (CLI Runner)
+This subsystem orchestrates the lifecycle of deploying temporary slots, discovering routes, spidering endpoints, and persisting telemetry.
+*   **`src/apphosting/compare/suite.ts`**: The main execution orchestrator for the CLI command. It processes test matrices, locks concurrency slots, and schedules deployment loops.
+*   **`src/apphosting/compare/lifecycle.ts`**: Handles programmatic interactions with the Firebase CLI (`firebase deploy`) and cleanups.
+*   **`src/apphosting/compare/discover.ts`**: Parses local build metadata (such as the Next.js pre-render manifest) to compile an initial target route list.
+*   **`src/apphosting/compare/crawler.ts`**: A recursive spider crawler that walks paths on live deployments to auto-discover dynamic routes.
+*   **`src/apphosting/compare/cache.ts`**: Manages reading and writing serialization records under the `.compare-cache/` workspace directory.
+*   **`src/apphosting/compare/compare.ts`**: Core parity engine containing the text similarity scoring (Levenshtein/Jaccard), status check rules, and header variation classifications.
+*   **`src/apphosting/compare/distance.ts`**: Algorithmic library for calculating string similarity distances.
+
 ---
+
 
 ## Instructions for AI Agents
