+++ src/apphosting/compare/suite.ts
@@ -1,3 +1,4 @@
+// [Subsystem: Deployment, Routing Discovery & Crawling Orchestrator]
 import * as path from "path";
 import * as fs from "fs-extra";
 import * as apphosting from "../../gcp/apphosting";
